+++ greeter/greet.py
@@ -2,15 +2,26 @@
 A simple program to greet people to the class.
 """
 import inspect
+import pathlib
+import importlib.util
 
-from greeter import people
 
-
-def find_people():
+def import_module_from_path(filepath):
     """
-    Returns the functions prefixed with `get`.
+    Import a module from given path.
+    https://stackoverflow.com/a/67692/5159551
     """
-    functions = inspect.getmembers(people, predicate=inspect.isfunction)
+    spec = importlib.util.spec_from_file_location('person', filepath)
+    person_module = importlib.util.module_from_spec(spec)
+    spec.loader.exec_module(person_module)
+    return person_module
+
+
+def find_people(person_module):
+    """
+    Returns the functions prefixed with `get` from a module.
+    """
+    functions = inspect.getmembers(person_module, predicate=inspect.isfunction)
     people_functions = filter(lambda x: x[0].startswith('get'), functions)
 
     persons = []
@@ -32,7 +43,12 @@ def greet_people():
     """
     Greet a participants of this project.
     """
-    people_list = find_people()
+    people_list = []
+    for path_to_module in pathlib.Path(__file__).parent.glob('people/*.py'):
+        person_module = import_module_from_path(path_to_module)
+        module_people_list = find_people(person_module)
+        people_list.extend(module_people_list)
+
     print(get_welcome_message(people_list))
