+++ greeter/person.py
@@ -19,16 +19,3 @@ class Person:
 
     def __repr__(self) -> str:
         return '{}, {}'.format(self.lastname.capitalize(), self.name.capitalize())
-
-
-def get_metin_senturk() -> Person:
-    """
-    Information on instructor. 
-    """
-    return Person(
-        name='metin',
-        lastname='senturk',
-        date_of_birth=date(1989, 9, 6),
-        educational_history=['Bilkent University', "Saint Peter's University"],
-        skills=dict(python='*****', sql='**')
-    )
